unify ViewModel tests
