+++ viewmodel/src/jsMain/kotlin/com/hoc081098/kmp/viewmodel/internal/AtomicBoolean.js.kt
