+++ viewmodel/src/jsTest/kotlin/com/hoc081098/kmp/viewmodel/utils/runBlockInNewThread.js.kt
