+++ viewmodel/src/nativeMain/kotlin/com/hoc081098/kmp/viewmodel/internal/AtomicBoolean.native.kt
