+++ viewmodel/src/nativeTest/kotlin/com/hoc081098/kmp/viewmodel/SynchronizedTest.native.kt
