+++ viewmodel/src/nonAndroidTest/kotlin/com/hoc081098/kmp/viewmodel/LockableObjectTest.nonAndroid.kt
