+++ viewmodel/src/nonAndroidTest/kotlin/com/hoc081098/kmp/viewmodel/utils/TestAtomicBoolean.nonAndroid.kt
@@ -1,6 +1,8 @@
 package com.hoc081098.kmp.viewmodel.utils
 
+import com.hoc081098.kmp.viewmodel.internal.AtomicBoolean
+
 internal actual class TestAtomicBoolean actual constructor(value: Boolean) {
-  private val ref = com.hoc081098.kmp.viewmodel.internal.AtomicBoolean(value)
+  private val ref = AtomicBoolean(value)
   actual var value: Boolean by ref::value
 }
